PROJ1615: Optimize away as much extraction as possible.
For all of the software where we extract a tarball (and patch it)
We do this every time, even if the result would be identical to
what we currently have.

This change will only perform the extraction and patching
iff the result would be different.

Extract is now delayed out of the package (lua) loading stage.

During the extraction stage:

- A new file is created in the work directory (.extraction.info.new)
  This file contains the type, name, (extra info) and hash of the tar
   or patch file.
- It is then compared to .extraction.info
  iff these files differ,
   - The source extraction process is run
   - The package is marked to indicate that the source code has changed
- The new .extraction.info.new is moved over .extraction.info

The fetch() implementation has been modified to mark the package
 as having the source code updated (irrespective of change) if
 any files are copied/linked/cloned into the work directory

Signed-off-by: Mark Wutzke <mark.wutzke@alliedtelesis.co.nz>