PROJ1615: Optimize away building.
Continuing on from optimizing away extraction.
This patch will prevent a package from being rebuilt
iff,
- the package lua file is identical
and
- the staging and install tarballs exist
and
- the order and value of checked feature values is identical

As with extraction optimization, a new file is created
.build.info (.build.info.new will exist while the build is occuring)

There is also a new fetch mechanism, 'copyfile'.
This helps buildsys to understand when a file is being copied,
instead of an entire directory.
Buildsys will record copied files in the extraction info.

Signed-off-by: Scott Parlane <scott.parlane@alliedtelesis.co.nz>