MAINT: Allow buildsys to optimize away building entirely
Given the frequency of continuous builds, and the stability of
a given build output versus the build.info file,
We can make use of the continuous build output to improve the
clean build times.

The continuous builds will need to invoke update_cache.py
to populate the cache, and the users will need to specify
PACKAGES="--ff http://awpbuild.atlnz.lc/buildsys_cache/"
to activate this feature.

Note that --ff (fetch from) is equal too --cache-server

Signed-off-by: Andrew McGregor <andrew.mcgregor@alliedtelesis.co.nz>