PROJ1164: Pass build options through to invoke build.
This change moves invokebuild so that it is now executed during the building
stage and not the pre prepossessing stage. This allows for faster builds
as two builds are run at the same time. Running two builds at the same time
does cause higher cpu usage.

The invoked build is now passing the build options from the parent buildsys
task. This allows functionality like the "--clean" to be handled in the
child build as well as any package filters.

After this change invokebuild will not require the buildsys path and the
path to the buildsys++ executable from lua.

Reviewed-by: Mark Wutzke <mark.wutzke@alliedtelesis.co.nz>
Reviewed-by: Scott Parlane <scott.parlane@alliedtelesis.co.nz>