bd:fetch drop '-d' from 'copyfile'
The -d option causes symbolic links to be de-referenced when copying.
The downside of this is that we can't have multiple build targets re-use
the same config file by linking from one to the other. There appears no
reason for keeping the -d option. If the de-referencing behaviour is
needed then the "copy" method is still available.

Change-Id: I6749eecce93071a75433520bf5b91da1d786e742
Signed-off-by: Scott Parlane <scott.parlane@alliedtelesis.co.nz>