Allow packages to hash their output to improve rebuild time.
Packages that extract a small low-change selection of data
from a bigger high-change source data set can now help their
'depend()'ers by 'hashoutput()'ing.

'hashoutput()' will hash the entire output of the package and
all 'depend()'ing packages will use the hash of this file,
where they would normally use the hash of the build.info file.

The package doing 'hashoutput()' will have to rebuild if
it's source data or a package it 'depend()'s on changes,
but that wont force anything that 'depend()'s on it to rebuild
if the output hasn't changed.

Change-Id: I7b80f523f89fbac4fc9a8a283836f9dcf4ea0de7
Reviewed-by: Mark Wutzke <mark.wutzke@alliedtelesis.co.nz>
Reviewed-by: Chris Packham <chris.packham@alliedtelesis.co.nz>
Reviewed-by: John Collis <john.collis@alliedtelesis.co.nz>
Signed-off-by: Scott Parlane <scott.parlane@alliedtelesis.co.nz>