MAINT: Use flock for "autoreconf" and "configure"
It happens that when ATLAS libraries built, "configure" is corrupted for example
for CFC400 builds or when multiple builds are compiled at the same time.

Now added "flock SRC_DIR" for "autoreconf" and "configure" commands which will
assure only one of "autoreconf" or "configure" to run in the same source code
directory at a time.

Change-Id: I7ca1fb9588a1d3858480cc2b62193c0c980470da
Reviewed-by: Scott Parlane <scott.parlane@alliedtelesis.co.nz>

Signed-off-by: Scott Parlane <scott.parlane@alliedtelesis.co.nz>