Add package overlays support
Allow multiple directories to be search for package files.
the package/ directory is always searched first, then
each of the arguments to --overlay are searched in order.
bd:fetch() operations that are package-base relative will
be relative to whichever directory the package file was
found in. (this means you can't directly override a specific
file that that a package fetches without overriding the
entire package directory)

New argument: --overlay PATH
where PATH is the overlay to add to the package file search.
for best results, use a relative path.

Change-Id: I3cb90bf379687b5ee6958c9b9a20f3073b0cff91
Reviewed-by: Chris Packham <chris.packham@alliedtelesis.co.nz>
Reviewed-by: Mark Wutzke <mark.wutzke@alliedtelesis.co.nz>
Reviewed-by: John Collis <john.collis@alliedtelesis.co.nz>