BLDSYS-11: Use a seperate lua interpreter for each package lua file
Currently all packages are processed by the same lua interpreter
which leaves variables around that can be accidentally re-used by
other packages and cause random breakages.

This patch moves the Lua object from the World class to the Package
class. This causes a new Lua interpreter to be used for each package.

Change-Id: Iab9bd5d9f90dfce57d96d1102a6045ed19821847
Reviewed-by: Scott Parlane <scott.parlane@alliedtelesis.co.nz>
Reviewed-by: Joshua Scott <joshua.scott@alliedtelesis.co.nz>
Reviewed-by: John Collis <john.collis@alliedtelesis.co.nz>
Reviewed-by: Chris Packham <chris.packham@alliedtelesis.co.nz>

Signed-off-by: Scott Parlane <scott.parlane@alliedtelesis.co.nz>