BLDSYS-12: Intercept require() statements
Currently, files that are "require()"d in the lua script are not part of the
dependencies, which means changes made in the require() file will not trigger
the package to rebuild.

The require() files are now added into packages' .build.info, so packages will
rebuild when it detect changes in the require() file.

Change-Id: I0c44a4052d7a5389fa602c5e7a925dd4db0aa25a
Reviewed-by: Mahoda Ratnayaka <mahoda.ratnayaka@alliedtelesis.co.nz>
Reviewed-by: Scott Parlane <scott.parlane@alliedtelesis.co.nz>
Signed-off-by: Scott Parlane <scott.parlane@alliedtelesis.co.nz>