BLDSYS: Respect an absolute path when copying a file
An issue was seen where if copying a file with:

    bd:fetch('/some/file', 'copyfile')

Instead of copying the file with the path left unchanged, the working
directory would be added to the beginning of the path. This resulted
in files not being found due to having two copies of the working
directory in the filepath.

Change-Id: Ie0cbf3e6a4543a06b1e7f76a5190aa7711d476ae
Reviewed-by: Ridge Kennedy <ridge.kennedy@alliedtelesis.co.nz>
Reviewed-by: Scott Parlane <scott.parlane@alliedtelesis.co.nz>
Signed-off-by: Scott Parlane <scott.parlane@alliedtelesis.co.nz>