BLDSYS-71: Defer extraction until build time
Firstly, take the package code that was controlling extraction,
and move it to the Extraction class.

Then remove the code in World that was calling extract directly,
and the code that was causing extract() to be recursive.
This is replaced by a call to Package::extract() just before
it does prepareBuildInfo()

Change-Id: Ia8cf1c6048bc6bf4e4e774133b19639e77bbc355
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: Ridge Kennedy <ridge.kennedy@alliedtelesis.co.nz>