BLDSYS-49: Fix issues with hashing of files.
The compressed file HASH() function was never being run,
this causes the hash of any downloaded file to be ignored.

Also, whenever hash_file is called, it assumes the current
directory is the buildsys dir, however there are some places
where this is possibly not the case.

Add the buildsys.lua file to make bd:shell and friends work,
also extend the test to try to download a file

Change-Id: Ied4fbbd63fe26853a4f2106ff0c27e3374a702b6
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: Ridge Kennedy <ridge.kennedy@alliedtelesis.co.nz>