Allow fetch to saved downloaded files with a different file name
Some tarballs downloaded from the upstream source location have poor
naming (i.e. "v0.4.0.tar.gz"). To allow the contents of the downloaded
tarball to be easier identified add another argument to li_bd_fetch()
that allows the user to specify a custom name to save the downloaded
tarball as.

Change-Id: I7eadeaf648d0ba38407d2c026c0c2ba24e32e260
Reviewed-by: Scott Parlane <scott.parlane@alliedtelesis.co.nz>
Reviewed-by: Ridge Kennedy <ridge.kennedy@alliedtelesis.co.nz>
Signed-off-by: Scott Parlane <scott.parlane@alliedtelesis.co.nz>