fetch: git, get hash from package if available
Don't fetch git repos just to get the hash if it is present
in the package Digest file.
This is only suitable for git tags.

Modifies GitExtractionUnit to do some of the steps that
were being done in fetch in the init function instead.
Also modifies it's dependencies so they can be created
with no arguments.

Don't add the GitExtractionUnit as a fetch unit

Change-Id: I0f35c4faf0cd4eb46e5fb555f16af44e5356199b
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: Ridge Kennedy <ridge.kennedy@alliedtelesis.co.nz>
Signed-off-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>