MAINT: Make it possible to change the remote url
If the git url of a package ever changed, it would never
be updated and therefore no new code would ever be fetched.
This left the user to figure out what was wrong and probably
delete the source directory (or not use this feature).
Instead, make sure origin points to the url provided,
and force a fetch if the url changes.

Change-Id: I2830e1b9fa4cd0f6fba61e4f4e764c2d314c7924
Reviewed-by: Ridge Kennedy <ridge.kennedy@alliedtelesis.co.nz>
Signed-off-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>