package: Search all the overlays for files
This allows fetching in an overlay where
the files to be fetched is stored in a
different overlay. Which also means it's
possible to override some of the files
without copying the entire package.

Also, switch require to use this rather
than it's own implementation that didn't
handle all possible cases.

Change-Id: I91ab33923a1e465a0021e69b948bda861fa83971
Reviewed-by: Ridge Kennedy <ridge.kennedy@alliedtelesis.co.nz>
Signed-off-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>