package: remove the staging directory after building
Staging directories account for a large portion
of the disk space used by the build.
Almost always, we don't need them after the
package has been built.
Provide a way for packages to mark that the
contents of the staging directory need to stay
and delete the staging directory on a successful
build if this flag isn't set.

Change-Id: Ic25f42e0aa2a2988f531340916f722d1017ec2ea
Reviewed-by: Chris Packham <chris.packham@alliedtelesis.co.nz>
Signed-off-by: Ridge Kennedy <ridge.kennedy@alliedtelesis.co.nz>