fetch: lock downloads while in progress
Some downloads are used by different packages. Sometimes
both packages will try to download the file at the same
time, which leads to one of them seeing an empty file
and aborting because the hash doesn't match the expectation.

Create a list of all the files that are required, and
have to take and release a lock before hashing the file.
While the lock is held, download the file if required.
This prevents other packages from hashing an incomplete file.
The locks are per-file to allow parallel downloading to occur.

Change-Id: I2a77b99cb74363b8badd1eb22dd8b33acf20238a
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Signed-off-by: Ridge Kennedy <ridge.kennedy@alliedtelesis.co.nz>