info: Don't include the overlay path in info files
Previously both build.info and extraction.info could contain
the path to the overlay that actually contains the file used.
This resulted in differences between the build.info files
between building the same package in the overlay and building
it from another repo, also different repos using the same
overlay(s) may or may not have gotten the same build.info
depending on the path to the overlay.

Instead, strip the overlays path from the build.info. The
hash will still ensure rebuilding if the contents of the file
changes.

The build cache used by --cache-server is equally impacted by
the difference in build.info contents. This change allows
the build cache to be populated by the overlay and used by
a repo that refers to that overlay, as would be expected.

Equally applies to patch files used in the extraction.info,
so fix that too.

Change-Id: Iba0cdb05bfc000b472119cae7749c11a04150f98
Reviewed-by: Ridge Kennedy <ridge.kennedy@alliedtelesis.co.nz>
Signed-off-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>