fetch deps: Add listedonly to remove the need for intercept()
Instead of teaching every package that it is part of a special group,
allow the package that needs them to be special to do achieve what it wants.

The intercept() marker allows packages to mark themselves as the boundary
of what to pickup during fetch { method = 'deps' }. This is good for
packages that produce an image of everything they depend on.
intercept() was also being used to limit the scope of what is included
by packages that make images of their direct dependencies but not other
pacakges. This leaves intercept() in packages that don't make images
and some confusion over when it should be there.
Instead of this, allow packages that want to only make images of their
dependencies to say so, and not force intercept() in packages that don't care.

Change-Id: I1275f62c09e7ccce9b4f5eadda4adb91ead13409
Reviewed-by: Ridge Kennedy <ridge.kennedy@alliedtelesis.co.nz>
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>