Fix #include ordering
clang-format puts the includes alphabetically, "" before <>. This breaks
compilation as the include of buildsys.h is sometimes required before other
includes (for types etc). Fix this properly by including buildsys.h in
luainterface.h.

Also switch to including buildsys.h in quotes and specify the path in the
includes to please cpplint.

Change-Id: Icc8168fcbaaf1255c36d5a841b1009cd3282d7c2
Reviewed-by: Scott Parlane <scott.parlane@alliedtelesis.co.nz>
Reviewed-by: Ridge Kennedy <ridge.kennedy@alliedtelesis.co.nz>
Signed-off-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>