Enable -Wextra compiler flag
Fix any issues raised by this flag. This finds a number of functions
with unused parameters. Rework the code to avoid unused parameters.
In the two functions that we cannot remove the unused parameter (as
they must have the correct function signature to be registered as a
LUA callback) simply remove the parameter names to make the C++ compiler
happy. Note that there is an issue in cpplint where it sees this as
a C style cast so simply add the NOLINT comment in these two places.

Change-Id: If86310be02622239cbcbdc3bad618b0de7fa2b59
Reviewed-by: Scott Parlane <scott.parlane@alliedtelesis.co.nz>
Signed-off-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>