package: Improve dependency set creation
Currently we iterate through every package and its dependencies and
rely on the unordered_set container to ensure that a package only
appears in the set once. Even though this is correct it is
inefficient to iterate again through a package that is already in
the set. Therefore we now check if the package is already in the
set before iterating through its dependencies.

Change-Id: Iaf41f2be4ce005617be35af9785f951b498ad175
Reviewed-by: Mark Wutzke <mark.wutzke@alliedtelesis.co.nz>
Reviewed-by: Scott Parlane <scott.parlane@alliedtelesis.co.nz>
Signed-off-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>