package: Ensure --quietly and --clean do not conflict with each other
Currently when --quietly and --clean are used together the log file for
the package (via --quietly) is removed (via --clean) meaning there are
no logs for the package. The easiest and chosen solution for this problem
is to move the package log from "/output/<namespace>/<package>/work/build.log"
to "/output/<namespace>/<package>/build.log" as --clean only removes the
work directory.

Change-Id: Iff67f3db6665fd9982b9e81788d6496d2e13b27e
Reviewed-by: Scott Parlane <scott.parlane@alliedtelesis.co.nz>
Signed-off-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>