buildinfo: Remove direct hash knowledge
Remove the need for the BuildDescription class to hash files (pass
the hash in instead). This is done for consistency and to simplify
the class.

Note that the buildinfo hash is already stored on the package and
passed in. For performance reasons we wouldn't want to change this
to make the BuildDescription class hash the file again each time a
package depends on a given package.

Change-Id: I5f0e9a7635abb813834375d75aa88774d857fab5
Reviewed-by: Scott Parlane <scott.parlane@alliedtelesis.co.nz>
Signed-off-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>