package: hold the build lock while testing the built-ness too
The was_built flag is used to decide if something that was already
built needs to be built again, because another package needs it
to have been built locally.
If multiple packages try to cause a package to build locally at
the same time, the package will end up being built multiple times
because each instance will see was_built as false, and then never
check it again.
These packages are going to get held up anyway, so we should hold
the lock while checking the built-ness of the package.

Change-Id: I85ab562c69d3d5adb6690291b0dada4209ded8b0
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Signed-off-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>