main: Install signal handler for SIGINT, SIGHUP and SIGQUIT
Since commit 36faad8 ("packagecmd: Add pty support") the ability to
kill buildsys++ from the terminal with CTRL+C stopped working. It is
assumed that the underlying implementation of 'forkpty' somehow changes
the default signal handling of the process such that it is no longer
killed.

Explicitly installing signal handling for these signals ensures that
buildsys++ can now be killed.

Change-Id: I95e517b46722f292abf9ddc0bea6a4c0f7e8098d
Reviewed-by: Mark Wutzke <mark.wutzke@alliedtelesis.co.nz>
Reviewed-by: Scott Parlane <scott.parlane@alliedtelesis.co.nz>
Signed-off-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>