packagecmd: Reset signals and unblock all
It seems that in glibc 2.29 development the way that the system call
is implemented changed such that it now has a side-effect of changing
the behaviour for some signals and some of them now become masked.
Once we fork, in the child process reset all signals to the defaults,
and unblock all signals so that processes we start get a state closer
to how they expect to have been started.

Change-Id: I046f1e903951f91c8e2cb831d930b196661e01af
Reviewed-by: Mark Wutzke <mark.wutzke@alliedtelesis.co.nz>
Signed-off-by: Ridge Kennedy <ridge.kennedy@alliedtelesis.co.nz>