world: when keep-going is used, log which package(s) failed to build
Using the --keep-going option can mean that the package which failed
scrolls entirely off the screen, which makes it difficult to figure
out which package is failing.
So, in the case that keep-going is used, once all the running packages
are complete, let the ones that failed log that fact again.

Change-Id: Ic486e8d9e7caba941701adc3e899e9410b084b7d
Reviewed-by: Chris Packham <chris.packham@alliedtelesis.co.nz>
Signed-off-by: Ridge Kennedy <ridge.kennedy@alliedtelesis.co.nz>