intercept: allow intercepting the staging directory as well
Previously it was only possible to intercept the install contents.
We now have some control files that are passed around via staging
that we want to stop progressing further.
Also, this is useful to allow packages to replace files in staging
if they need to modify them (i.e. modify header files).

Change-Id: I372d00335762a4e6e80c0337ca61be405f57808c
Reviewed-by: Callum Sinclair <callum.sinclair@alliedtelesis.co.nz>
Signed-off-by: Ridge Kennedy <ridge.kennedy@alliedtelesis.co.nz>