Add --fetch-only option
Add the ability to just perform the fetch operation. This skips
extraction and any possible application of patches, as well as
skipping the building of the package.
This is being added to add with detection of source tarballs whose
locations have moved, or where the hashes of tarballs in our cache
mismatch the upstream in some way.

Change-Id: I14aa1795205680049b1e2c844d1a6a13a6f47451
Reviewed-by: Scott Parlane <scott.parlane@alliedtelesis.co.nz>
Reviewed-by: Ridge Kennedy <ridge.kennedy@alliedtelesis.co.nz>
Signed-off-by: Ridge Kennedy <ridge.kennedy@alliedtelesis.co.nz>